API/requirements.txt
pydantic~=2.10.6
fastapi~=0.115.7
SQLAlchemy~=2.0.37
asyncpg~=0.30.0
greenlet~=3.1.1
alembic~=1.14.1
psycopg2-binary~=2.9.10
python-dotenv~=1.0.1
PyJWT~=2.10.1
python-multipart~=0.0.20
uvicorn~=0.34.0

faiss-cpu
langchain==0.3.15
numpy==1.25.1
python-dotenv==1.0.1
sentence_transformers==3.3.1
torch==2.0.1
transformers==4.48.1
yandex_cloud_ml_sdk==0.2.0
rank-bm25==0.2.2